backend/requirements.txt
Flask
mysql-connector-python
deepface
opencv-python
numpy
tensorflow
keras
torch
torchvision
PyCryptodome
Pillow
